+++ packages/thumbprint-react/utils/use-focus-trap.ts
@@ -15,9 +15,18 @@ function toggleTrap(trap: FocusTrap, isActive: boolean): void {
  *
  * @param element  The element to trap focus inside of
  * @param isActive Whether or not to activate the trap
+ * @param initialFocus The initial element inside the trap to focus
  */
-export default function useFocusTrap(element: HTMLElement | null, isActive = false): void {
+export default function useFocusTrap(
+    element: HTMLElement | null,
+    isActive = false,
+    initialFocus?: HTMLElement | null,
+): void {
     const [trap, setTrap] = useState<FocusTrap>();
+
+    const initialFocusElement = initialFocus || element;
+
+    console.log('focusing on', initialFocusElement ? initialFocusElement.type : undefined);
 
     useEffect((): (() => void) => {
         // If we've already created a trap, toggle it based on the isActive status
@@ -30,7 +39,7 @@ export default function useFocusTrap(element: HTMLElement | null, isActive = fal
                 clickOutsideDeactivates: true,
                 // Set initial focus to the modal wrapper itself instead of focusing on the first
                 // focusable element by default
-                initialFocus: element,
+                initialFocus: initialFocusElement,
             });
             setTrap(newTrap);
             // And toggle it based on isActive status
@@ -43,5 +52,5 @@ export default function useFocusTrap(element: HTMLElement | null, isActive = fal
                 trap.deactivate();
             }
         };
-    }, [element, isActive, trap]);
+    }, [element, isActive, trap, initialFocusElement]);
 }